requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
coverage==7.6.4
Django==5.1.2
django-extensions==3.2.3
django-ninja==1.3.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-decouple==3.8
ruff==0.7.1
sqlparse==0.5.1
typing_extensions==4.12.2
